Add asset twig extension if not loaded
