+++ tests/project/app/Resources/views/asset.pug
@@ -1,0 +1,2 @@
+extend /layout/layout.pug
+    p=asset("/foo")
